refine knowledgebase
